Move selectors to their own package
