+++ pkg/streamingpromql/operators/selectors/instant_vector_selector.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Prometheus Authors
 
-package operators
+package selectors
 
 import (
 	"context"
