+++ pkg/streamingpromql/operators/selectors/selector.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package operators
+package selectors
 
 import (
 	"context"
